requirements.txt
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
statsmodels==0.14.1
xgboost==2.0.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
azure-storage-blob==12.19.0
azure-identity==1.15.0
python-dotenv==1.0.0
jupyter==1.0.0
pytest==7.4.3
imbalanced-learn==0.11.0
shap==0.44.0
joblib==1.3.2
openpyxl==3.1.2
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
httpx==0.26.0
pyyaml==6.0.1
